requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-multipart==0.0.20
jinja2==3.1.6
httpx==0.28.1
Pillow==11.3.0
numpy==2.0.2
opencv-python-headless==4.13.0.92
scikit-image==0.24.0
scipy==1.13.1
